convert onMapReady to postFrameCallback remove map controller from options
